Add files via upload
I made this table from the giant unique ids sheet, but simplified it slightly. I wanted to make sure that for each Name id, I had a birth date (if known, unknown for immigrants), stop date (either death or disappearance), and sex (M or F). This will help me tremendously when cleaning and coding the avoids and grooming data. I think that the table is updated as of December 2022. I deleted any ids with a stop date prior to 2017, since I foresee most of my analyses will be from 2018 onward.